+++ pkg/network/tracer/connection/util/utils.go
@@ -1,0 +1,75 @@
+package util
+
+import (
+	"os"
+
+	manager "github.com/DataDog/ebpf-manager"
+	"github.com/cilium/ebpf/asm"
+
+	"github.com/DataDog/datadog-agent/pkg/ebpf"
+	ebpftelemetry "github.com/DataDog/datadog-agent/pkg/ebpf/telemetry"
+	"github.com/DataDog/datadog-agent/pkg/network/config"
+	"github.com/DataDog/datadog-agent/pkg/network/ebpf/probes"
+	"github.com/DataDog/datadog-agent/pkg/security/utils"
+	"github.com/DataDog/datadog-agent/pkg/util/log"
+)
+
+// ComputeDefaultClosedConnRingBufferSize is the default buffer size of the ring buffer for closed connection events.
+// Must be a power of 2 and a multiple of the page size
+func ComputeDefaultClosedConnRingBufferSize() int {
+	numCPUs, err := utils.NumCPU()
+	pageSize := os.Getpagesize()
+	if err != nil {
+		numCPUs = 1
+	}
+	if numCPUs <= 16 {
+		return 8 * 8 * pageSize
+	}
+	return 8 * 16 * pageSize
+}
+
+// computeDefaultClosedConnPerfBufferSize is the default buffer size of the perf buffer for closed connection events.
+// Must be a multiple of the page size
+func computeDefaultClosedConnPerfBufferSize() int {
+	return 8 * os.Getpagesize()
+}
+
+// SetupClosedConnHandler sets up the closed connection event handler
+func SetupClosedConnHandler(connCloseEventHandler ebpf.EventHandler, mgr *ebpftelemetry.Manager, cfg *config.Config) {
+	switch handler := connCloseEventHandler.(type) {
+	case *ebpf.RingBufferHandler:
+		options := manager.RingBufferOptions{
+			RecordGetter:     handler.RecordGetter,
+			RecordHandler:    handler.RecordHandler,
+			TelemetryEnabled: cfg.InternalTelemetryEnabled,
+			// RingBufferSize is not used yet by the manager, we use a map editor to set it in the tracer
+			RingBufferSize: ComputeDefaultClosedConnRingBufferSize(),
+		}
+		rb := &manager.RingBuffer{
+			Map:               manager.Map{Name: probes.ConnCloseEventMap},
+			RingBufferOptions: options,
+		}
+
+		mgr.RingBuffers = []*manager.RingBuffer{rb}
+		ebpftelemetry.ReportRingBufferTelemetry(rb)
+	case *ebpf.PerfHandler:
+		pm := &manager.PerfMap{
+			Map: manager.Map{Name: probes.ConnCloseEventMap},
+			PerfMapOptions: manager.PerfMapOptions{
+				PerfRingBufferSize: computeDefaultClosedConnPerfBufferSize(),
+				Watermark:          1,
+				RecordHandler:      handler.RecordHandler,
+				LostHandler:        handler.LostHandler,
+				RecordGetter:       handler.RecordGetter,
+				TelemetryEnabled:   cfg.InternalTelemetryEnabled,
+			},
+		}
+		mgr.PerfMaps = []*manager.PerfMap{pm}
+		ebpftelemetry.ReportPerfMapTelemetry(pm)
+		helperCallRemover := ebpf.NewHelperCallRemover(asm.FnRingbufOutput)
+		err := helperCallRemover.BeforeInit(mgr.Manager, nil)
+		if err != nil {
+			log.Error("Failed to remove helper calls from eBPF programs: ", err)
+		}
+	}
+}
